認証機能追加

+++ index.mjs
@@ -1,20 +1,23 @@
 import express from "express";
 import env from "dotenv";
 import connectDB from "./utils/database.mjs";
+import auth from "./utils/auth.mjs";
+import jwt from "jsonwebtoken";
 import { ItemModel, UserModel } from "./utils/schemaModels.mjs";
 env.config();
 
+const SECRET_KEY = process.env.SECRET_KEY || "exmaple-secret";
 const PORT = process.env.PORT || 5050;
 const app = express();
 app.use(express.urlencoded({ extended: true }));
 app.use(express.json());
 
 // ITEM function
-
 // Create Item
-app.post("/item/create", async (req, res) => {
+app.post("/item/create", auth, async (req, res) => {
   try {
     await connectDB();
+    console.log(req.body);
     await ItemModel.create(req.body);
     return res.status(200).json("Item create success");
   } catch (err) {
@@ -46,9 +49,13 @@ app.get("/item/:id", async (req, res) => {
   }
 });
 // Update Item
-app.put("/item/update/:id", async (req, res) => {
+app.put("/item/update/:id", auth, async (req, res) => {
   try {
     await connectDB();
+    const singleItem = await ItemModel.findById(req.params.id);
+    if (singleItem.email !== req.body.email) {
+      throw new Error();
+    }
     await ItemModel.findByIdAndUpdate(req.params.id, req.body);
     return res.status(200).json({ message: "Item update success" });
   } catch (err) {
@@ -56,9 +63,13 @@ app.put("/item/update/:id", async (req, res) => {
   }
 });
 // Delete Item
-app.delete("/item/delete/:id", async (req, res) => {
+app.delete("/item/delete/:id", auth, async (req, res) => {
   try {
     await connectDB();
+    const targetItem = await ItemModel.findById(req.params.id);
+    if (targetItem.email !== req.body.email) {
+      throw new Error();
+    }
     await ItemModel.findByIdAndDelete(req.params.id);
     return res.status(200).json({ message: "Item delete success" });
   } catch (err) {
@@ -86,7 +97,12 @@ app.post("/user/login", async (req, res) => {
       return res.status(400).json({ message: "User not found" });
     }
     if (user.password === req.body.password) {
-      return res.status(200).json({ message: "User login success" });
+      // passworc success
+      const payload = {
+        email: req.body.email,
+      };
+      const token = jwt.sign(payload, SECRET_KEY, { expiresIn: "1d" });
+      return res.status(200).json({ message: "User login success", token });
     } else {
       return res.status(400).json({ message: "Password is incorrect" });
     }
